Change filetype to html.handlebars
With this change, some tools can have a better understanding that handlebars are an extension of html. 

Tools like Onivim uses `html.*` to start a HTML language server, with this change now that works nicely.

Another point is that when using `vim-polyglot`, which includes vim-mustache-handlebars, it also sets the filetype as `html.handlebars` and now I can disable `handlebars` from `vim-polyglot` and this plugin will work.